+++ src/rogallo/screens/main/screen.py
@@ -74,7 +74,6 @@ from ...commands import (
     StripeLinks,
     ToggleANSIEscapeSequenceHandling,
     ToggleCosyLinkNumbers,
-    ToggleEmojiRemoval,
     ToggleLinkNumbers,
     ToggleSidePanel,
     ToggleView,
@@ -246,7 +245,6 @@ class Main(EnhancedScreen[None]):
         StripeLinks,
         ToggleANSIEscapeSequenceHandling,
         ToggleCosyLinkNumbers,
-        ToggleEmojiRemoval,
         ToggleLinkNumbers,
         ToggleSidePanel,
         ToggleView,
@@ -372,7 +370,6 @@ class Main(EnhancedScreen[None]):
         self._viewer.stripe_links = config.stripe_links
         self._viewer.with_link_numbers = config.with_link_jumps
         self._viewer.handle_ansi_escape_sequences = config.handle_ansi_escape_sequences
-        self._viewer.strip_emoji = config.strip_emoji
         self._viewer.cosy_link_numbers = config.cosy_link_jumps
         if self._arguments.command == "open" and (
             location := getattr(self._arguments, "location", None)
@@ -903,12 +900,6 @@ class Main(EnhancedScreen[None]):
             and location.root != location
         ):
             self.post_message(OpenLocation(location.root))
-
-    def action_toggle_emoji_removal_command(self) -> None:
-        """Toggle emoji removal."""
-        self._viewer.strip_emoji = not self._viewer.strip_emoji
-        with update_configuration() as config:
-            config.strip_emoji = self._viewer.strip_emoji
 
     def action_toggle_ansi_escape_sequence_handling_command(self) -> None:
         """Toggle ANSI escape sequence handling."""