+++ ChangeLog.md
@@ -29,6 +29,9 @@
 - **BREAKING CHANGE:** Moved `side_panel_visible`, `side_panel_on_right` and
   `side_panel_chosen_tab` out of configuration and turned them into UI state
   properties. ([#453](https://github.com/davep/rogallo/pull/453))
+- **BREAKING CHANGE:** Removed the `ToggleEmojiRemoval` command. This is now
+  controlled by a key binding specific to the viewer widget.
+  ([#453](https://github.com/davep/rogallo/pull/453))
 
 ## v2.4.0